Added documentation to various classes Changed Update to Updater
