+++ src/org/montclairrobotics/sprocket/utils/PID.java
@@ -50,7 +50,7 @@ public class PID implements Updatable{
 		this.minIn=minIn;
 		this.maxIn=maxIn;
 		setTarget();
-		Update.add(this);
+		Updater.add(this);
 	}
 
 	public void setPID(double P, double I, double D){
